app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
numpy==1.26.4
openai==1.30.1
python-dotenv==1.0.1
